create Authorization using jwt and create an action for deleting a post using api

+++ controllers/api/v1/posts_api.js
@@ -26,13 +26,17 @@ module.exports.index = async (req, res)=>{
 module.exports.destroy = async function(req, res){
     let post = await Post.findById(req.params.id);
     try{
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             await post.remove();
             await Comment.deleteMany({post: req.params.id});
             return res.json(200, {
                 message: 'Post and associated comments are deleted.'
             });
-        // }
+        }else{
+            return res.json(401, {
+                message: "Unauthorized!! You cannot delete this post"
+            })
+        }
     }catch(err){
         console.log("Hey error due ");
         console.log('*****', err);
